backend/requirements.txt
# Pathway with LLM docs
pathway[xpack-llm-docs]==0.21.4

# LLM related
litellm>=1.65.0
google-generativeai>=0.8.0

# Basic libraries
sentence-transformers>=2.6.0
torch>=2.2.0
transformers>=4.38.0

# Configuration management
pydantic-settings>=2.0.0
pyyaml>=6.0.0

# Required libraries
python-dotenv
requests>=2.32.2
numpy

# Suppress beartype warnings
beartype>=0.15.0